Created initial API class
Removed httparty - it seemed like it would be easier to use uri and
net/http instead as I was getting an error that wasn't easily resolved.
Got API call working with a predetermined word. Added JSON parsing to
turn the JSON into Ruby. Updated gem requirements in environment and
gemfile.