Modified API related functions to get the word that is entered
Word class now calls the fetch_word_details_from_API method in the API
class. CLI class calls the fetch_details method in the Word class -
currently it just returns a hash of all the details, but the next step
is modifying fetch_details to assign the details to the word's instance
attributes.