Finish first working version
Add methods for saving all the details of a word to the attributes of
the Word instance.
Build out method for user to choose what details they want and view them.
Add helper method to flatten detail arrays and remove nil values.
Add helper method to print the array values nicely without having too
much code repetition.
Remove CLI method for selecting which definition to learn about - for
most words there was not enough information for any one definition for
the program to be useful. From a user's perspective, I would rather see
all the synonyms for all meanings at once and ignore the ones that don't
work for me than have to keep looping through the program to check the
details for all the definitions.
Build out user interfact for CLI.
Add method to allow the user to get more details or look up another word
before exiting the program.
Change the flow of the CLI class so the call method only starts the
first method, and each method after calls the next method that should
follow it.